requirements.txt
python-ldap
python-gitlab